Luis

+++ src/app/component/accesorios/accesorios.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { ApiService } from '../../services/api.service';
 import { RouterLink } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-accesorios',
@@ -44,7 +45,26 @@ export class AccesoriosComponent {
 
 
   agregarProductoCarrito(producto:any){
-    this.dataCarritoLS.push(producto)
-    localStorage.setItem("carrito", JSON.stringify(this.dataCarritoLS))
+    Swal.fire({
+      title: "Estas seguro?",
+      text: "Deseas agregar el producto a tu compra",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Agregar al carrito"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.dataCarritoLS.push(producto)
+        localStorage.setItem("carrito", JSON.stringify(this.dataCarritoLS))
+        Swal.fire({
+          title: "Este producto fue agregado",
+          text: ("revisa el resumen de compra"),
+          icon: "success"
+        });
+      }
+    });
+
+
   }
 }
